requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
SQLAlchemy>=2.0.30
playwright>=1.44.0
playwright-stealth>=1.0.6
google-generativeai>=0.5.2
python-dotenv>=1.0.1
pydantic>=2.7.2
psycopg2-binary>=2.9.9
